+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Hi! This is where we document all notable changes, including bug fixes, enhancem
 ### Changed
 
 - Improved form design.
+- Updated RLS policies and front end to prevent "self-dealing": scholars who enrich themselves with tokens, without oversight.
 - Updated internal tooling for stability.
 
 ## 0.3.14 - 2026-05-10